Update using other Ada Jumpstart students
Tatsqui, I can't figure out how to leave a comment or message you, but your code was super helpful. It helped me realize to define options, create a conditional flow for candy choices to assign names and prices for the computer, and to create a conditional flow for calculating prices (including if the option was not one offered). Thank you.